Changement de int -> Float

+++ src/main/java/ulaval/glo2003/CriteriaMaxPrice.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class CriteriaMaxPrice implements Criteria{
-    int maxPrice;
+    float maxPrice;
 
-    public CriteriaMaxPrice(int p_maxPrice){
+    public CriteriaMaxPrice(float p_maxPrice){
         this.maxPrice = p_maxPrice;
     }
     @Override
